Address Daniel's comments and DRY a bit
